:ambulance:!HOTFIX: delete profiles app #10 #6 #2
one - to - one 필드는 연결된 모델이 생성될때 오브젝트가 생성되지
않으므로 put 메서드에 접근하여 프로필을 수정한다 할지라도 데이터베이스에
데이터가 생성되지 않은 치명적인 오류 발견,
따라서 User Model과 결합
