+++ B4games/settings.py
@@ -40,7 +40,6 @@ INSTALLED_APPS = [
     'rest_framework_simplejwt',
     'users',
     'posts',
-    'profiles',
     'comments',
     "corsheaders",
 ]
