+++ users/views.py
@@ -4,7 +4,8 @@ from rest_framework.response import Response
 from rest_framework.generics import get_object_or_404
 from .models import User
 from rest_framework_simplejwt.views import TokenObtainPairView
-from .serializers import (UserSerializer,ComtomTokenObtainPairSerializer,ReadUserSerializer)
+from .serializers import (UserSerializer,ComtomTokenObtainPairSerializer,ReadUserSerializer,
+                          ReadProfileSerializer,UpdateProfileSerializer,GetFollowInfoSerializer)
 from datetime import datetime
 
 # 회원 가입
@@ -61,3 +62,41 @@ class UserView(APIView):
 class CustomTokenObtainPairView(TokenObtainPairView):
     serializer_class = ComtomTokenObtainPairSerializer
 
+class UpdateProfileView(APIView):
+    # owner의 프로필 읽기 (public)
+    def get(self,request,user_id):
+        owner = get_object_or_404(User,id=user_id)
+        serializer = ReadProfileSerializer(owner)
+        return Response(serializer.data,status=status.HTTP_200_OK)
+
+    # owner를 팔로우,언팔로우
+    def post(self,request,user_id):
+        # 로그인 인증 테스트 필요
+        if not request.user.is_authenticated:
+            return Response("로그인이 필요합니다.", status=status.HTTP_401_UNAUTHORIZED)
+
+        owner = get_object_or_404(User,id=user_id)
+
+        if request.user == owner:
+            return Response("나 자신을 팔로우 할 수 없습니다.", status=status.HTTP_400_BAD_REQUEST)
+
+        if request.user in owner.followers.all():
+            owner.followers.remove(request.user)
+            return Response("unfollow",status=status.HTTP_200_OK)
+        else:
+            owner.followers.add(request.user)
+            return Response("follow",status=status.HTTP_200_OK)
+
+    # owner가 자신의 프로필을 수정
+    def put(self,request,user_id):
+        owner = get_object_or_404(User, id=user_id)
+        if request.user == owner:
+            serializer = UpdateProfileSerializer(owner, data=request.data, partial=True)
+            if serializer.is_valid():
+                serializer.save()
+                update_profile_info = ReadProfileSerializer(owner)
+                return Response(update_profile_info.data, status=status.HTTP_200_OK)
+            else:
+                return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+        else:
+            return Response({"error": "권한이 없습니다."}, status=status.HTTP_400_BAD_REQUEST)
